% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_IGN_5_data.R
\name{download_IGN_5_data}
\alias{download_IGN_5_data}
\title{Download IGN 5 mil layers}
\arguments{
\item{layer_name}{The specific layer you want to download (e.g., IGN_5:delimitacion2017_5k).}

\item{output_format}{The format in which to download the data (default is GeoJSON, application/json).}

\item{output_file}{The path where the downloaded spatial data will be saved (e.g., .gpkg, .shp).}
}
\description{
Download IGN 5 mil layers
}
\examples{

Select a layer from the list (replace with your chosen layer)

Replace with the actual layer name from the list

layer_name <- "IGN_5:delimitacion2017_5k"

Specify your desired output file format and name

output_file <- "IGN_5_delimitacion2017_5k.gpkg"

Download the selected layer

downloaded_file <- download_IGN_5_data(layer_name, output_file = output_file)



}
