% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_IGN_5k.R
\name{get_IGN_5k}
\alias{get_IGN_5k}
\title{Download IGN 1:5000 scale layers}
\usage{
get_IGN_5k(layer_name, output_file = output_file)
}
\arguments{
\item{layer_name}{The specific layer you want to download (e.g., IGN_5:delimitacion2017_5k).}

\item{output_format}{The format in which to download the data (default is GeoJSON, application/json).}

\item{output_file}{The path where the downloaded spatial data will be saved (e.g., .gpkg, .shp).}
}
\value{
a file
}
\description{
This function allows downloading layers at a 1:5000 scale from
the Instituto Geográfico Nacional, Costa Rica, through
Sistema Nacional de Información Territorial de Costa Rica
 \href{https://www.snitcr.go.cr}{SNIT}
}
\examples{

Select a layer from the list (replace with your chosen layer)

Replace with the actual layer name from the list

layer_name <- "IGN_5:delimitacion2017_5k"

Specify your desired output file format and name

output_file <- "IGN_5_delimitacion2017_5k.gpkg"

Download the selected layer

get_IGN_5k(layer_name, output_file = output_file)



}
