% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_IGN_25k.R
\name{list_IGN_25k}
\alias{list_IGN_25k}
\title{List IGN 1:25000 scale layers}
\usage{
list_IGN_25k()
}
\value{
a list of the available layers
}
\description{
This function list the available layers at a 1:25000 scale from
the Instituto Geográfico Nacional, Costa Rica, through
Sistema Nacional de Información Territorial de Costa Rica
 \href{https://www.snitcr.go.cr}{SNIT}
}
\examples{

List available layers

capas <- list_IGN_25k()

}
