% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_IGN_5_layers.R
\name{list_IGN_5_layers}
\alias{list_IGN_5_layers}
\title{List IGN 5 mil layers}
\description{
This function list the available IGN 5 mil layers
}
\examples{

List available layers

capas <- list_IGN_5_layers()

}
