% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_IGN_5k.R
\name{list_IGN_5k}
\alias{list_IGN_5k}
\title{List IGN 5 mil layers}
\usage{
list_IGN_5k()
}
\description{
This function list the available layers at a 1:5000 scale from
the Instituto Geográfico Nacional, Costa Rica, through
Sistema Nacional de Información Territorial de Costa Rica
 \href{https://www.snitcr.go.cr}{SNIT}
}
\examples{

List available layers

capas <- list_IGN_5k()

}
