% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snitr-package.R
\docType{package}
\name{snitr-package}
\alias{snitr}
\alias{snitr-package}
\title{snitr: This package download data from snit (https://www.snitcr.go.cr)}
\description{
This package download data from snit (https://www.snitcr.go.cr)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ManuelSpinola/snitr}
  \item Report bugs at \url{https://github.com/ManuelSpinola/snitr/issues}
}

}
\author{
\strong{Maintainer}: Manuel Spínola \email{mspinola10@gmail.com} (\href{https://orcid.org/0000-0002-7839-1908}{ORCID})

}
\keyword{internal}
